Remove Nitrous.io
The new Nitrous Pro has a daily usage limit that cripples free accounts. Without a usable free account, those learning JS will be unable to use it free for educational purposes, and those trying to evaluate it for purchase will be unable to get a fair evaluation of its capabilities.